feat(major restructuring): implement controller and request handler classes, refactor routing
This update involves implementing new controller and request handler classes, along with refactoring
the routing system. The new controllers are responsible for managing distinct functionalities of the
application, enhancing modularity and clarity in the codebase. Request handlers are designed to
process incoming HTTP requests, ensuring they're directed to the appropriate controllers, while
handling validation and error management. The refactoring of the routing system aligns with these
new structures, improving the efficiency and clarity of URL routing and request processing. Overall,
these changes aim to make the application more robust, maintainable, and scalable, facilitating
easier updates and expansions in the future.
